# Fixes #30101 - Use delete_all for SubscriptionFacetPools association