Deleting a pool would issue individual SQL queries when cleaning up SubscriptionFacetPool objects:

```
2020-06-10T15:36:46 [D|sql|]   Katello::SubscriptionFacetPool Destroy (0.2ms)  DELETE FROM "katello_subscription_facet_pools" WHERE "katello_subscription_facet_pools"."id" = $1  [["id", 476743]] 
```

Using delete_all issues a single query:
```
2020-06-11T13:07:23 [D|sql|]   Katello::SubscriptionFacetPool Destroy (29.1ms)  DELETE FROM "katello_subscription_facet_pools" WHERE "katello_subscription_facet_pools"."pool_id" = $1  [["pool_id", 394]] 
```

In a real-world scenario this affects manifest import speed, and the speedup is significant:

before:
```
5: Actions::Candlepin::Owner::ImportProducts (success) [ 2351.84s / 2351.84s ]
```

after:
```
5: Actions::Candlepin::Owner::ImportProducts   (success)        [ 3.78s / 3.78s ]
```

In a simulated scenario I saw speedup like this for a single pool with 50000 SubscriptionFacetPools when calling `Katello::Pool.import_all`

before:
```
=> [#<Benchmark::Tms:0x00000000179f48b0 @cstime=0.0, @cutime=0.0, @label="", @real=14.058943261043169, @stime=0.767635, @total=10.296436, @utime=9.528801>]
```

after (29ms was spent deleting)
```
=> [#<Benchmark::Tms:0x000000001561f318 @cstime=0.0, @cutime=0.0, @label="", @real=0.12148274498758838, @stime=0.003206999999999738, @total=0.030070000000048225, @utime=0.026863000000048487>]
```

This is how I set up the scenario: https://gist.github.com/jturel/317e9f892fb19e84480880f1f674e74e