@jturel 2020-06-11
> would it make sense to have some utility function that you could use like:
> 
> ```
> Katello::Util.time("My Log message", :logger => my_logger) do 
>   Some Code
> end
> ```
> 
> which would log something like: "My Log Message took 3.3 seconds". I think it would reduce a lot of duplicated code (and could default to Rails.logger)

Nifty idea. I'll do it